requirements.txt
backports.shutil-get-terminal-size==1.0.0
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
Cython==0.29.9
decorator==4.4.0
docx==0.2.4
enum34==1.1.6
get==2019.4.13
idna==2.8
importlib==1.0.4
ipython==5.8.0
ipython-genutils==0.2.0
lxml==4.3.3
magickwand==0.2
nltk==3.4.1
numpy==1.16.4
pathlib2==2.3.3
pdf2image==1.5.4
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
post==2019.4.13
prompt-toolkit==1.0.16
ptyprocess==0.6.0
public==2019.4.13
Pygments==2.4.2
PyPDF2==1.26.0
pytesseract==0.2.6
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
scandir==1.10.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
traitlets==4.3.2
urllib3==1.25.3
Wand==0.5.4
wcwidth==0.1.7
win-unicode-console==0.5
